Move `Witness` to `Cardano.Read.Ledger.Tx.Witnesses`
