+++ lib/api/src/Cardano/Wallet/Api/Http/Server/Handlers/TxCBOR.hs
@@ -40,6 +40,9 @@ import Cardano.Read.Ledger.Tx.ReferenceInputs
 import Cardano.Read.Ledger.Tx.Validity
     ( getEraValidity
     )
+import Cardano.Read.Ledger.Tx.Witnesses
+    ( getEraWitnesses
+    )
 import Cardano.Wallet.Api.Http.Server.Error
     ( IsServerError (..)
     , apiError
@@ -66,9 +69,6 @@ import Cardano.Wallet.Read.Eras
     )
 import Cardano.Wallet.Read.Eras.EraFun
     ( applyEraFun
-    )
-import Cardano.Wallet.Read.Tx.Witnesses
-    ( getEraWitnesses
     )
 import Cardano.Wallet.Transaction
     ( TokenMapWithScripts
