+++ lib/read/cardano-wallet-read.cabal
@@ -63,6 +63,7 @@ library
     Cardano.Read.Ledger.Tx.ScriptValidity
     Cardano.Read.Ledger.Tx.Validity
     Cardano.Read.Ledger.Tx.Withdrawals
+    Cardano.Read.Ledger.Tx.Witnesses
     Cardano.Wallet.Read
     Cardano.Wallet.Read.Block
     Cardano.Wallet.Read.Block.BHeader
@@ -97,7 +98,6 @@ library
     Cardano.Wallet.Read.Tx.Inputs
     Cardano.Wallet.Read.Tx.TxId
     Cardano.Wallet.Read.Tx.TxIn
-    Cardano.Wallet.Read.Tx.Witnesses
 
   build-depends:
     , base
